requirements.txt
# WhisperX Transcription Service - Requirements
# Python 3.10 recommended

# =============================================================================
# Core ML/AI
# =============================================================================
whisperx
faster-whisper

# PyTorch (installed separately in Dockerfile with CUDA support)
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# DO NOT include torch/torchaudio here - they're pre-installed

# NLP / ML
transformers
pyannote.audio

# =============================================================================
# Audio Processing
# =============================================================================
librosa
soundfile
ffmpeg-python

# =============================================================================
# Translation & LLM
# =============================================================================
google-genai

# =============================================================================
# API Server
# =============================================================================
fastapi>=0.100.0
uvicorn[standard]
python-multipart

# =============================================================================
# Database
# =============================================================================
sqlalchemy[asyncio]>=2.0.0
asyncpg  # PostgreSQL async driver
passlib[bcrypt]  # Password hashing
python-jose[cryptography]  # JWT tokens

# =============================================================================
# Task Queue
# =============================================================================
celery[redis]>=5.3.0
redis>=4.0.0
flower  # Celery monitoring (optional)

# =============================================================================
# Document Generation
# =============================================================================
python-docx
openpyxl  # Excel generation for tasks.xlsx
weasyprint  # HTML to PDF for risk_brief.pdf

# =============================================================================
# Data Validation & Utils
# =============================================================================
pydantic>=2.0
pydantic-settings
python-dotenv
numpy
tqdm
psutil  # System monitoring for admin stats

# =============================================================================
# Security
# =============================================================================
slowapi>=0.1.9  # Rate limiting for FastAPI

# =============================================================================
# Development (optional)
# =============================================================================
pytest
pytest-asyncio
httpx
aiosqlite

